project/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
python-multipart==0.0.20
paddlepaddle==3.1.0
paddleocr==3.1.0
pillow==11.3.0
requests==2.32.3
aiofiles==24.1.0 